Moved exec_line_ops to its own file
