+++ monty.h
@@ -51,6 +51,7 @@ typedef struct global_state
 extern state_t global;
 
 /* opcode handlers */
+void exec_line_ops(char *line, size_t line_num);
 void push(stack_t **, unsigned int);
 void pall(stack_t **, unsigned int);
 void pint(stack_t **, unsigned int);
